pyproject.toml
[tool.poetry]
name = "cover-jobby"
description = "Inputs list of books and downloads thumbnail images"
authors = ["Matt Cossin"]
version = "0.5.0"


[tool.poetry.dependencies]
python = "=3.11.4"
requests = "^2.31.0"
python-slugify = "^8.0.1"
pillow = "^10.0.0"
poetry = "^1.5.1"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
commitizen = "^3.5.2"

[tool.commitizen]
version = "0.6.0"
update_changelog_on_bump = true
version_files = ["pyproject.toml:^version"]
